pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monoai"
version = "0.0.12"
description = "Pacchetto MonoAI sviluppato da ProfessionAI"
authors = [
  { name = "ProfessionAI Tech", email = "tech@profession.ai" }
]
readme = "README.md"
requires-python = ">=3.8"

# 📦 Dipendenze di runtime
dependencies = [
    "litellm",
    "duckduckgo-search",
    "tavily-python",
    "typing-extensions",
    "xmltodict",
    "pytest",
    "beautifulsoup4",
    "selenium",
    "python-whois",
    "chromadb",
    "python-docx",
    "docx2txt",
    "PyPDF2"
]

[project.urls]
Homepage = "https://github.com/Profession-AI/MONOAI"
